docker_course_udemy/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
autopep8==2.0.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-extensions==3.2.1
django-grappelli==3.0.4
django-tinymce==3.5.0
djangorestframework==3.14.0
dnspython==2.3.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.0
parso==0.8.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
ptpython==3.0.22
pycodestyle==2.10.0
Pygments==2.14.0
pymongo==4.3.3
pytz==2022.7
PyYAML==6.0
requests==2.28.2
rich==13.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
